incorporated Anant's comments fixed the endpoint of monitoring in config
