+++ src/jcsclient/utils.py
@@ -48,7 +48,7 @@ def load_service(service):
     try:
         current_mod = get_dir_name(__file__)
         return importlib.import_module('.' + service,
-                                      package=current_mod)
+                                       package=current_mod)
     except ImportError as ie:
         raise exception.ServiceNotFound(service)
 
@@ -158,44 +158,6 @@ def dash_to_camelcase(keyword):
         keyword += part
     return keyword
 
-def push_monitoring_indexed_params(params, key, vals):
-    """Populate the params dict for list of vals
-    dimensions will be converted to dimensions.member.1.{}
-
-    param params: dictionary to populate
-
-    param key: key to be used in the dictionary
-
-    param vals: list of values to be saved in the dict
-
-    return: Nothing
-    """
-    idx = 0
-    for val in vals:
-        idx += 1
-        elements = val
-        key_index = key + '.member.' + str(idx)
-        # This is for cases like --filter 'Name=xyz,Values=abc'
-        elements = val.split(',')
-        if len(elements) == 1 and val.find('=') == -1:
-            params[key_index] = val
-            continue
-        for element in elements:
-            if element.find('=') != -1:
-                parts = element.split('=')
-                if len(parts) != 2:
-                    msg = 'Unsupported value ' + element + 'given in request.'
-                    raise ValueError(msg)
-                element_key, element_val = parts[0], parts[1]
-                if element_key == 'Values':
-                    element_key = element_key[:-1] + "." + str(idx)
-                updated_key = key_index + '.' + element_key
-                params[updated_key] = element_val
-            else:
-                msg = 'Bad request syntax. Please see help for valid request.'
-                raise ValueError(msg)
-
-
 def push_indexed_params(params, key, vals):
     """Populate the params dict for list of vals
 
@@ -241,6 +203,7 @@ def push_indexed_params(params, key, vals):
                 msg = 'Bad request syntax. Please see help for valid request.'
                 raise ValueError(msg)
 
+
 def get_protocol_and_host(url):
     """Validate a given url and extract the protocol and
        host from given url.
@@ -260,26 +223,6 @@ def get_protocol_and_host(url):
         return (None, None)
     else:
         return (url_parts.group(1), url_parts.group(2))
-
-
-def populate_monitoring_params_from_args(params, args):
-    """After the argparser has processed args, populate the
-       params dict, processing the given args for mitoring.
-
-       param params: a dict to save the processed args
-
-       param args: Namespace object where args are saved
-
-       returns: None
-    """
-    if not isinstance(args, dict):
-        args = vars(args)
-    for arg in args:
-        key = underscore_to_camelcase(arg)
-        if isinstance(args[arg], list):
-            push_monitoring_indexed_params(params, key, args[arg])
-        elif args[arg]:
-            params[key] = args[arg]
 
 
 def populate_params_from_cli_args(params, args):
